teste

+++ models/trusted/google_sheets/tb_discounts.sql
@@ -1,5 +1,5 @@
 select
     _ROW as id,
     order_value as vl_order,
-    cast(discount as numeric(1,2)) as vl_discount
+    cast(discount as double(1,2)) as vl_discount
 from {{ source('google_sheets', 'googlesheets_discounts') }}
